Index - Moved all Index code to one place: index.js
